com.hammed/requirements.txt
click==8.1.7
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==2.3.7
gunicorn
flask_sqlalchemy
Flask-WTF
Flask-Limiter
Flask-Login==0.6.3
Flask-Mail
python-dotenv